invalid product, error added while deleteing product

+++ backend/controllers/product.controller.js
@@ -38,13 +38,17 @@ export const createProduct = async(req, res) => {
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;  //This take id from the URL.
     
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false, message:"invalid product id"});
+    }
+
    try {
         await Product.findByIdAndDelete(id); //finds the product in MongoDB and deletes it.
         res.status(200).json({ success: true, message: "Product deleted"});
 
     } catch (error) {
         console.log("error in dlting product:",error.message);
-        res.status(404).json({ success: false, message: "Product not found"});
+        res.status(500).json({ success: false, message: "server error"});
         
     }
 };
